requirements.txt
streamlit
seaborn
matplotlib
pandas
numpy
africastalking
google-generativeai
reportlab
pillow
scikit-learn
